+++ src/provider/utils/crud/create.ts
@@ -5,10 +5,7 @@ import { DBCreateItemParams, DynamodbExecutor } from '../dynamoDB';
 import { ItemExpression } from '../expressions';
 import { getConditionParams } from '../conditions';
 
-export interface CreateItemParams<
-  Entity,
-  PKs extends StringKey<Entity> | unknown = unknown,
-> {
+export interface CreateParams<Entity, PKs extends StringKey<Entity> | unknown = unknown> {
   /**
    * Dynamodb Table
    */
@@ -39,7 +36,7 @@ export class ItemCreator extends DynamodbExecutor {
     item,
     table,
     conditions,
-  }: CreateItemParams<Entity>): DBCreateItemParams['input'] {
+  }: CreateParams<Entity>): DBCreateItemParams['input'] {
     return {
       TableName: table,
 
@@ -49,7 +46,7 @@ export class ItemCreator extends DynamodbExecutor {
     };
   }
 
-  async create<Entity>(params: CreateItemParams<Entity>): Promise<Entity> {
+  async create<Entity>(params: CreateParams<Entity>): Promise<Entity> {
     await this._insertItem(this.getCreateParams(params));
 
     return params.item;