+++ src/singleTable/implementation.ts
@@ -10,7 +10,7 @@ import {
   ListItemTypeResult,
   SingleTableBatchGetParams,
   SingleTableQueryParams,
-  SingleTableCreateItemParams,
+  SingleTableCreateParams,
   SingleTableGetParams,
   SingleTableTransactionParams,
   SingleTableUpdateParams,
@@ -54,7 +54,7 @@ export class SingleTable<SingleParams extends SingleTableParams> {
   }
 
   async create<Entity>(
-    params: SingleTableCreateItemParams<Entity, SingleParams>,
+    params: SingleTableCreateParams<Entity, SingleParams>,
   ): Promise<Entity> {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     return this.methods.create(params as any);